+++ src/x/xcb/convert.rs
@@ -176,13 +176,13 @@ impl From<&ClientConfig> for Vec<(u16, u32)> {
                 vec![
                     (xcb::CONFIG_WINDOW_X as u16, geom.x as u32),
                     (xcb::CONFIG_WINDOW_Y as u16, geom.y as u32),
-                    (xcb::CONFIG_WINDOW_HEIGHT as u16, geom.height),
-                    (xcb::CONFIG_WINDOW_WIDTH as u16, geom.width),
+                    (xcb::CONFIG_WINDOW_HEIGHT as u16, geom.height as u32),
+                    (xcb::CONFIG_WINDOW_WIDTH as u16, geom.width as u32),
                 ]
             }
             Resize {h, w} => vec![
-                (xcb::CONFIG_WINDOW_HEIGHT as u16, *h),
-                (xcb::CONFIG_WINDOW_WIDTH as u16, *w),
+                (xcb::CONFIG_WINDOW_HEIGHT as u16, *h as u32),
+                (xcb::CONFIG_WINDOW_WIDTH as u16, *w as u32),
             ],
             Move {x, y} => vec![
                 (xcb::CONFIG_WINDOW_X as u16, *x as u32),
